Moved firmware to a separate header. Changed quaternion to float after determining the fixed point format to be Q1.30.

+++ dmp.h
@@ -5,7 +5,7 @@
 
 #include "mpu6050.h"
 
-int32_t dmp_quaternion(uint8_t index);
+float dmp_quaternion(uint8_t index);
 
 void DMPInit(void);
 enum MPU6050Error DMPReadFIFO(void);
